Example dataset for genetic association meta-analysis of type 2 diabetes (T2D)

This dataset is an example data frame containing variables needed for two-sample Mendelian randomization or genetic association meta-analysis. It is intended to demonstrate the workflow of functions in this package (e.g., mr_MEITests), and users can use it as a template to prepare their own data.

Format

A data frame with rows corresponding to genetic variants (e.g., SNPs) and the following 6 numeric variables:

SNP

SNP

A1

A1

A2

A2

b.exp

Numeric vector of beta coefficients for genetic associations with the exposure variable. Corresponds to the parameter g_hat.

se.exp

Standard errors of b.exp. Corresponds to the parameter gse.

b.out

Numeric vector of beta coefficients for genetic associations with the outcome variable. Corresponds to the parameter G_hat.

se.out

Standard errors of b.out. Corresponds to the parameter Gse.

eaf.exp

Effect allele frequency in the exposure sample. Corresponds to the parameter EAF.exp.

eaf.out

Effect allele frequency in the outcome sample. Corresponds to the parameter EAF.out.

Details

This dataset is simulated and does not reflect real genetic association results. All variables are numeric vectors of the same length (i.e., number of genetic variants). When using analysis functions from this package, users can directly pass the columns of this data frame to the respective arguments.

Source

Simulated data, for illustrative and testing purposes only.

Examples

# View the first few rows
head(Meta_T2D)
#>              SNP A1 A2      b.exp        b.out     se.exp     se.out  eaf.exp
#> 3793  rs10031339  G  A  0.0198571  0.034420100 0.00638728 0.00995775 0.898844
#> 6389  rs10053125  C  T -0.0248670 -0.015468400 0.00454600 0.00739199 0.768022
#> 9527   rs1007766  G  A  0.0256093  0.016119900 0.00384711 0.00584018 0.482695
#> 9753   rs1007938  A  G -0.0176519 -0.011377400 0.00392348 0.00612014 0.593741
#> 10484 rs10086120  T  C -0.0103310  0.000819522 0.00386927 0.00589890 0.559912
#> 15528 rs10122199  A  G  0.0159693  0.005079630 0.00440304 0.00651100 0.736177
#>         eaf.out
#> 3793  0.9058045
#> 6389  0.8069720
#> 9527  0.4909820
#> 9753  0.6500740
#> 10484 0.5651840
#> 15528 0.7169790

# Example usage with a hypothetical mr_analysis() function
if (FALSE) { # \dontrun{
mr_MEITests(g_hat = Meta_T2D$b.exp,
            gse = Meta_T2D$se.exp,
            G_hat = Meta_T2D$b.out,
            Gse = Meta_T2D$se.out,
            EAF.exp = Meta_T2D$eaf.exp,
            EAF.out = Meta_T2D$eaf.out,
            eta=0.5,
            sel.pthr=5e-5)
} # }