The Modified Egger Intercept (MEI) Tests for Detecting Horizontal Pleiotropy in Mendelian Randomization

The modified Egger intercept (MEI) tests are developed to detect horizontal pleiotropy in two-sample summary-data Mendelian randomization.

mr_MEITests(
  g_hat,
  gse,
  G_hat,
  Gse,
  EAF.exp,
  EAF.out,
  eta = 0.5,
  sel.pthr = 5e-05
)

Arguments

g_hat

A numeric vector of beta-coefficient values for genetic associations with the exposure variable.

gse

The standard errors associated with the beta-coefficients g_hat.

G_hat

A numeric vector of beta-coefficient values for genetic associations with the outcome variable.

Gse

The standard errors associated with the beta-coefficients G_hat.

EAF.exp

A numeric vector of effect allele frequency in the exposure sample.

EAF.out

A numeric vector of effect allele frequency in the outcome sample.

eta

A parameter representing the standard deviation of the pseudo IV-exposure effect. By default, eta=0.5.

sel.pthr

The significance level threshold for IV selection. By default, sel.pthr=5e-5.

Value

The output from the function is a MEITests object containing:

Zvalue_major

Z-value from the MEI test under the major allele referenced coding scheme.

Pvalue_major

P-value associated with Zvalue_major.

Zvalue_normal

Z-value from the MEI test under the normal allele referenced coding scheme.

Pvalue_normal

P-value associated with Zvalue_normal.

Pvalue_combine

The P-value from the combined MEI test statistics obtained under major and normal allele referenced coding schemes.

SNPsNum

The number of SNPs after IV selection.

sel.pthr

The significance level threshold for IV selection.

Details

The modified Egger intercept (MEI) tests are developed to detect horizontal pleiotropy in two-sample Mendelian randomization with summary statistics, i.e., an exposure dataset (with IV-exposure effect g_hat and standard error gse) and an outcome dataset (with IV-outcome effect G_hat and standard error Gse).

We recommend that the input dataset be pre-processed with filtering using the rerandomization threshold approach, harmonisation, and clumping. eta and sel.pthr should be consistent with the parameters used during pre-filtering.

The MEI test under a specific allele coding scheme is constructed by deriving a bias-corrected Egger intercept estimator along with its variance under the null hypothesis of absence of both directional and correlated pleiotropy, leveraging the rerandomized inverse variance weighted (RIVW) estimator (see Ma et al., 2023). Under the null hypothesis, the test statistic follows a asymptotic standard normal distribution.

The MEI test statistic is computed under two allele coding schemes, along with a combined version: (i) major allele referenced coding; (ii) normal allele referenced coding, where the normal allele from the exposure GWAS is used as the reference; and (iii) a combined version, which aggregates the two test statistics under the major and normal allele schemes using the maximum Z-value approach.

EAF.exp and EAF.out are used to recode the summary data into the major allele referenced coding scheme. (1) If both EAF.exp and EAF.out > 0.5, the IV-exposure and IV-outcome effects are flipped. (2) If one of EAF.exp and EAF.out is missing, the SNP is flipped if the available value > 0.5. (3) In all other cases, the original allele coding is retained.

Examples

attach(Meta_T2D)
mr_MEITests(g_hat=b.exp,
            gse=se.exp,
            G_hat=b.out,
            Gse=se.out,
            EAF.exp=eaf.exp,
            EAF.out=eaf.out,
            eta=0.5,
            sel.pthr=5e-5)
#> Modified Egger intercept tests
#> 
#> IV selection threshold: 5e-05 
#> Number of variants : 343 
#> Null hypothesis: absence of both directional and correlated pleiotropy
#> ------------------------------------------------------------------
#>                           Method Z-value  P-value
#>   Major-allele-referenced coding  -0.093    0.926
#>  Normal-allele-referenced coding   3.922 8.78e-05
#>      Maximum Z-value combination         1.76e-04
#> ------------------------------------------------------------------
detach(Meta_T2D)